restructured header files for consistent presentation of c interface
moved all c++ specific implementation details, as well as references to any STL header
to their respective compilation units. created a seperate compilation unit for pthread
as to allow for forward declaring the necessary structs in a consistent c fashion, while
implementing it in c++ in the unit. this allows for seperation of header use for c
without pulling in any STL header, as this will lead to compiler errors on windows.
for details checkout https://github.com/ofiwg/libfabric/issues/7041#issuecomment-914839351
or the STL commit https://github.com/microsoft/STL/pull/2148
